collectionViewer Patch
This version is most updated version of collectionViewer. It builds a list that includes golden cards as well. Remember that you build the logic for the draft in the old version of this script, so if you download it, it will copy over all your code. Maybe you could move all the logic to it's own script?